+++ examples/basic/modules/_common/always/meta.nix
@@ -1,0 +1,3 @@
+{
+  before = [ "desktop.example" ];
+}