+++ examples/basic/modules/server/demo/meta.nix
@@ -1,0 +1,4 @@
+{
+  after = [ "development.demo" ];
+  wants = [ "_common.always" ];
+}